+++ FluidNC/src/Configuration/ParserHandler.h
@@ -23,13 +23,13 @@ namespace Configuration {
 
             // On entry, the token is for the section that invoked us.
             // We will handle following nodes with indents greater than entryIndent
-            int entryIndent = _parser._token._indent;
+            uint8_t entryIndent = _parser._token._indent;
             log_parser_verbose("Entered section " << name << " at indent " << entryIndent);
 
             // The next token controls what we do next.  If thisIndent is greater
             // than entryIndent, there are some subordinate tokens.
             _parser.Tokenize();
-            int thisIndent = _parser._token._indent;
+            uint8_t thisIndent = _parser._token._indent;
             log_parser_verbose("thisIndent " << _parser.key() << " " << thisIndent);
 
             // If thisIndent <= entryIndent, the section is empty - there are
@@ -39,7 +39,7 @@ namespace Configuration {
                 // If thisIndent > entryIndent, the new token is the first token within
                 // this section so we process tokens at the same level as thisIndent.
                 for (; _parser._token._indent >= thisIndent; _parser.Tokenize()) {
-                    log_parser_verbose(" KEY " << _parser.key() << " state " << int(_parser._token._state) << " indent "
+                    log_parser_verbose(" KEY " << _parser.key() << " state " << uint32_t(_parser._token._state) << " indent "
                                                << _parser._token._indent);
                     if (_parser._token._indent > thisIndent) {
                         log_error("Skipping key " << _parser.key() << " indent " << _parser._token._indent << " this indent " << thisIndent);
@@ -94,7 +94,7 @@ namespace Configuration {
             }
         }
 
-        void item(const char* name, int& value, const EnumItem* e) override {
+        void item(const char* name, uint32_t& value, const EnumItem* e) override {
             if (_parser.is(name)) {
                 value = _parser.enumValue(e);
             }
