+++ FluidNC/src/Kinematics/ParallelDelta.h
@@ -35,7 +35,7 @@ namespace Kinematics {
         virtual void init_position() override;
         //bool canHome(AxisMask& axisMask) override;
         bool cartesian_to_motors(float* target, plan_line_data_t* pl_data, float* position) override;
-        void motors_to_cartesian(float* cartesian, float* motors, int n_axis) override;
+        void motors_to_cartesian(float* cartesian, float* motors, uint8_t n_axis) override;
         bool transform_cartesian_to_motors(float* motors, float* cartesian) override;
         //bool soft_limit_error_exists(float* cartesian) override;
         bool         kinematics_homing(AxisMask& axisMask) override;
