+++ FluidNC/src/NutsBolts.cpp
@@ -15,7 +15,7 @@
 
 const int MAX_INT_DIGITS = 8;  // Maximum number of digits in int32 (and float)
 
-static float uint_to_float(uint32_t intval, int exp) {
+static float uint_to_float(uint32_t intval, int8_t exp) {
     float fval = (float)intval;
     // Apply decimal. Should perform no more than two floating point multiplications for the
     // expected range of E0 to E-4.
